Validation for deletion + softdeletes for CompleteLevel.php

+++ app/Actions/Jetstream/DeleteUser.php
@@ -3,6 +3,7 @@
 namespace App\Actions\Jetstream;
 
 use Laravel\Jetstream\Contracts\DeletesUsers;
+use Symfony\Component\HttpFoundation\Response;
 
 class DeleteUser implements DeletesUsers
 {
@@ -14,6 +15,10 @@ class DeleteUser implements DeletesUsers
      */
     public function delete($user)
     {
+        if ($errMsg = $user->userIsNotDeletable()) {
+            die($errMsg);
+        }
+
         $user->deleteProfilePhoto();
         $user->tokens->each->delete();
         $user->delete();
